\name{aovbayes}
\alias{aovbayes}
\title{Interactive panel ANOVA classic, non parametric and bayesian}
\usage{
aovbayes()
}
\description{
Interactive panel to visualize and develop one-way analysis of variance models, from the classical, non-parametric and Bayesian approach.
}
\examples{
aovbayes()
}
